+++ lib/checks/index.js
@@ -1,5 +1,5 @@
 const canAddLikeTermPolynomialNodes = require('./canAddLikeTermPolynomialNodes');
-const canDistribute = require('./canDistribute');
+const isPolynomialToConstantPower = require('./isPolynomialToConstantPower');
 const canMultiplyLikeTermConstantNodes = require('./canMultiplyLikeTermConstantNodes');
 const canMultiplyLikeTermPolynomialNodes = require('./canMultiplyLikeTermPolynomialNodes');
 const canRearrangeCoefficient = require('./canRearrangeCoefficient');
@@ -17,5 +17,5 @@ module.exports = {
   hasUnsupportedNodes,
   isQuadratic,
   resolvesToConstant,
-  canDistribute,
+  isPolynomialToConstantPower,
 };
