+++ lib/simplifyExpression/distributeSearch/index.js
@@ -19,7 +19,7 @@ function distribute(node) {
   if (Node.Type.isUnaryMinus(node)) {
     return distributeUnaryMinus(node);
   }
-  else if (checks.canDistribute(node)) {
+  else if (checks.isPolynomialToConstantPower(node)) {
     const [base, exponent] = node.args;
 
     let newNode;
@@ -29,9 +29,8 @@ function distribute(node) {
       newNode = Node.Creator.operator('*', Array(times).fill(base));
       return distributeAndSimplifyOperationNode(newNode);
     }
-    else {
-      return Node.Status.noChange(node);
-    }
+    
+    return Node.Status.noChange(node);
   }
   else if (Node.Type.isOperator(node)) {
     return distributeAndSimplifyOperationNode(node);
