+++ test/solveEquation/solveEquation.test.js
@@ -114,3 +114,16 @@ describe('constant comparison support', function () {
   ];
   tests.forEach(t => testSolveConstantEquation(t[0], t[1], t[2]));
 });
+
+function testEquationError(equationString, debug=false) {
+  it(equationString + ' throws error', (done) => {
+    assert.throws(() => solveEquation(equationString, debug),Error);
+    done();
+  });
+}
+
+describe('solveEquation errors', function() {
+  const tests = [
+  ];
+  tests.forEach(t => testEquationError(t[0], t[1]));
+});
